# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "contact": {
            "email": "pablomartinez123@hotmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/customer": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "operationId": "36d5a836e102529237aae81bf088183a",
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "operationId": "e788a12c4fbd7937feab964d2e133089",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/customer/{customer_id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "operationId": "560bd1ac2edcdc21a73566f9ac141a7c",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "description": "ID of customer to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Customer"
                ],
                "operationId": "5fe5ee0c796b5d10e1f94ccc1c2ee76f",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "description": "ID of customer to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customer"
                ],
                "operationId": "7900213a7179e0baf6cd15a8104ba570",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "description": "ID of customer to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payment-method": {
            "get": {
                "tags": [
                    "Payment Method"
                ],
                "operationId": "fab55bad7b89ba71ec0d18c3bfbee413",
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payment Method"
                ],
                "operationId": "e52b30d32703b0bd6cd25dab95dc20c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentMethod"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payment-method/{payment_method_id}": {
            "get": {
                "tags": [
                    "Payment Method"
                ],
                "operationId": "68843946c568ef4294d3becbc3a925a4",
                "parameters": [
                    {
                        "name": "payment_method_id",
                        "in": "path",
                        "description": "ID of payment method to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Payment Method"
                ],
                "operationId": "9e1b2a7f4b1c82ed21e14622dff4209d",
                "parameters": [
                    {
                        "name": "payment_method_id",
                        "in": "path",
                        "description": "ID of payment method to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Payment Method"
                ],
                "operationId": "17ee0c048fb0c297296ee4ad7ef4fa4f",
                "parameters": [
                    {
                        "name": "payment_method_id",
                        "in": "path",
                        "description": "ID of payment method to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Customer": {
                "title": "Customer",
                "required": [
                    "first_name",
                    "last_name",
                    "email"
                ],
                "properties": {
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PaymentMethod": {
                "title": "Payment Method",
                "required": [
                    "customer_id",
                    "name"
                ],
                "properties": {
                    "customer_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "external_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    }
}